feat: move getServiceVersion to common parsing
