+++ src/openApi/common/parser/service.ts
@@ -3,4 +3,6 @@
  * This basically removes any "v" prefix from the version string.
  * @param version
  */
-export const getServiceVersion = (version = '1.0'): string => String(version).replace(/^v/gi, '');
+export function getServiceVersion(version = '1.0'): string {
+    return String(version).replace(/^v/gi, '');
+}
