--- src/openApi/v2/parser/getServiceVersion.spec.ts
@@ -1,11 +1,0 @@
-import { describe, expect, it } from 'vitest';
-
-import { getServiceVersion } from './getServiceVersion';
-
-describe('getServiceVersion', () => {
-    it('should produce correct result', () => {
-        expect(getServiceVersion('1.0')).toEqual('1.0');
-        expect(getServiceVersion('v1.0')).toEqual('1.0');
-        expect(getServiceVersion('V1.0')).toEqual('1.0');
-    });
-});
