+++ src/openApi/v3/parser/__tests__/service.spec.ts
@@ -1,6 +1,6 @@
 import { describe, expect, it } from 'vitest';
 
-import { getServiceName, getServiceVersion } from '../service';
+import { getServiceName } from '../service';
 
 describe('getServiceName', () => {
     it('should produce correct result', () => {
@@ -12,13 +12,5 @@ describe('getServiceName', () => {
         expect(getServiceName('$fooBar')).toEqual('FooBar');
         expect(getServiceName('123fooBar')).toEqual('FooBar');
         expect(getServiceName('non-ascii-æøåÆØÅöôêÊ字符串')).toEqual('NonAsciiÆøåÆøÅöôêÊ字符串');
-    });
-});
-
-describe('getServiceVersion', () => {
-    it('should produce correct result', () => {
-        expect(getServiceVersion('1.0')).toEqual('1.0');
-        expect(getServiceVersion('v1.0')).toEqual('1.0');
-        expect(getServiceVersion('V1.0')).toEqual('1.0');
     });
 });
