+++ src/openApi/v3/parser/service.ts
@@ -12,10 +12,3 @@ export const getServiceName = (value: string): string => {
     const clean = sanitizeServiceName(value).trim();
     return camelCase(clean, { pascalCase: true });
 };
-
-/**
- * Convert the service version to 'normal' version.
- * This basically removes any "v" prefix from the version string.
- * @param version
- */
-export const getServiceVersion = (version = '1.0'): string => String(version).replace(/^v/gi, '');
